settings: 拆出 AuthorizedSitesController + AccountSettingsController
settings-controller.js 3077 → 43 行（仅持有 4 个 sub + 4 个委托方法）。
剩余的两个 sub-controller：

- AuthorizedSitesController（272 行）：授权站点列表/详情/UCAN 会话/撤销/
  清除全部授权。从主 class 拆出，独立注入 wallet。
- AccountSettingsController（239 行）：修改密码 modal/重置钱包 modal/「清除全部
  授权」按钮。为避免 sub-controller 互相依赖，「清除全部授权」按钮通过
  SettingsController 注入的 onClearAllAuthorizations 回调委托给 sites sub。

popup-controller 调用面零改动（loadAuthorizedSites/loadBackupSyncSettings/
loadMpcSettings/loadMpcSessions + bindEvents 全部保留）。
83+8=91 测试全绿、tsc 0 错误。UI 行为需 Chrome 实测（授权站点的列表/详情、
改密 modal、重置 modal、清除全部授权）。
